refactor: split stacked and static table into separate files
