+++ src/Table.js
@@ -1,0 +1,21 @@
+import { StackedTable } from './Table/StackedTable.js'
+import { StaticTable } from './Table/StaticTable.js'
+import { TableBody } from './Table/TableBody.js'
+import { TableCell } from './Table/TableCell.js'
+import { TableCellHead } from './Table/TableCellHead.js'
+import { TableFoot } from './Table/TableFoot.js'
+import { TableHead } from './Table/TableHead.js'
+import { TableRow } from './Table/TableRow.js'
+import { TableRowHead } from './Table/TableRowHead.js'
+
+export {
+    StaticTable,
+    StackedTable,
+    TableRow,
+    TableRowHead,
+    TableHead,
+    TableBody,
+    TableFoot,
+    TableCell,
+    TableCellHead,
+}
