+++ src/Table/TableCell.js
@@ -2,35 +2,13 @@ import React, { Fragment } from 'react'
 import css from 'styled-jsx/css'
 import propTypes from 'prop-types'
 
-import { Consumer } from './tableContext'
-import { TableCellText } from './TableCellText'
-import { colors } from '../theme'
-
-const tableCellStyles = css`
-    td {
-        border-bottom: 1px solid ${colors.grey300};
-        padding: 0 12px;
-        font-size: 14px;
-        line-height: 18px;
-    }
-
-    div {
-        min-height: 45px;
-    }
-
-    :global(tfooter) div {
-        min-height: 36px;
-    }
-`
-
-const tableCellStylesResponsive = css`
-    @media (max-width: 768px) {
-        td {
-            width: 100%;
-            display: block;
-        }
-    }
-`
+import { Consumer } from './tableContext.js'
+import { TableCellText } from './TableCellText.js'
+import {
+    tableCellStyles,
+    tableCellStylesStacked,
+    tableCellTitleStyleStacked,
+} from './styles.js'
 
 const TableCellStatic = ({ children, colSpan, rowSpan }) => (
     <td colSpan={colSpan} rowSpan={rowSpan}>
@@ -44,58 +22,26 @@ const ContentWithTitle = ({ title, children }) => (
         {title && <span className="title">{title}</span>}
         <span className="content">{children}</span>
 
-        <style jsx>{`
-            .title {
-                display: none;
-            }
-
-            .content {
-                display: block;
-            }
-
-            @media (max-width: 768px) {
-                :global(tbody) .title {
-                    display: block;
-                    white-space: normal;
-                    min-height: 24px;
-                    padding: 8px 0 0 0;
-                    white-space: nowrap;
-                    font-size: 13px;
-                    font-weight: normal;
-                    color: ${colors.grey800};
-                }
-
-                .content {
-                    display: block;
-                    padding: 0;
-                    min-height: 32px;
-                }
-
-                .content:first-child {
-                    padding-top: 8px;
-                    padding-bottom: 8px;
-                }
-            }
-        `}</style>
+        <style jsx>{tableCellTitleStyleStacked}</style>
     </Fragment>
 )
 
-const TableCellResponsive = ({ children, colSpan, rowSpan, title }) => (
+const TableCellStacked = ({ children, colSpan, rowSpan, title }) => (
     <td colSpan={colSpan} rowSpan={rowSpan}>
         <ContentWithTitle title={title}>{children}</ContentWithTitle>
 
         <style jsx>{tableCellStyles}</style>
-        <style jsx>{tableCellStylesResponsive}</style>
+        <style jsx>{tableCellStylesStacked}</style>
     </td>
 )
 
 export const TableCell = ({ children, colSpan, rowSpan, column }) => (
     <Consumer>
-        {({ responsiveLayout, headerLabels }) => {
-            const title = responsiveLayout ? headerLabels[column] : ''
+        {({ stackedLayout, headerLabels }) => {
+            const title = stackedLayout ? headerLabels[column] : ''
 
-            const TableCellComponent = responsiveLayout
-                ? TableCellResponsive
+            const TableCellComponent = stackedLayout
+                ? TableCellStacked
                 : TableCellStatic
 
             return (
