+++ src/Table/TableCellText.js
@@ -2,26 +2,8 @@ import React from 'react'
 import propTypes from 'prop-types'
 import css from 'styled-jsx/css'
 
-import { Consumer } from './tableContext'
-
-const tableCellTextStyles = css`
-    span {
-        display: inline-block;
-        padding: 9px 0;
-    }
-
-    :global(tbody) span {
-        padding: 13px 0;
-    }
-`
-
-const tableCellTextStylesResponsive = css`
-    @media (max-width: 768px) {
-        :global(tbody) span {
-            padding: 3px 0;
-        }
-    }
-`
+import { Consumer } from './tableContext.js'
+import { tableCellTextStyles, tableCellTextStylesStacked } from './styles.js'
 
 const TableCellTextStatic = ({ children }) => (
     <span>
@@ -30,19 +12,19 @@ const TableCellTextStatic = ({ children }) => (
     </span>
 )
 
-const TableCellTextResponsive = ({ children }) => (
+const TableCellTextStacked = ({ children }) => (
     <span>
         {children}
         <style jsx>{tableCellTextStyles}</style>
-        <style jsx>{tableCellTextStylesResponsive}</style>
+        <style jsx>{tableCellTextStylesStacked}</style>
     </span>
 )
 
 export const TableCellText = ({ children }) => (
     <Consumer>
-        {({ responsiveLayout }) => {
-            const TableCellTextComponent = responsiveLayout
-                ? TableCellTextResponsive
+        {({ stackedLayout }) => {
+            const TableCellTextComponent = stackedLayout
+                ? TableCellTextStacked
                 : TableCellTextStatic
 
             return <TableCellTextComponent children={children} />
